refactor: move auth components to appropriate route folder
