+++ app/auth/_components/Socials.tsx
@@ -1,6 +1,6 @@
 import { FaGithub } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
-import { Button } from "../ui/button";
+import { Button } from "../../../components/ui/button";
 import { signIn } from "next-auth/react";
 import { useSearchParams } from "next/navigation";
